Add files via upload
More results of the template matching are shown in Figures A.28,A.29,A.30 and A.31. Figures A.28 and A.29 show two experiments where templates of cores are matched against test images of bars. In both cases, no similarity is detected at all.
Figures A.30 and A.31 show two experiments where one bar template is matched against two test images, one showing a core and the second showing bars from another hydrogenerator with a very different design. In the case where the bar template is matched with a core (Figure A.30), no similarity is detected, as expected. When the bar template is matched against a test image of bars with a very different design (Figure A.31), some similarity is detected, as shown by the grey similarity matrix. This shows that the templates can find a match based on the context, even when the designs are significantly different.

Figure B.32 shows the simple model used for performance comparison in section 4. This model is a variation of the proposed model, except that there are no attention layers in this model.